adds dependency com.itextpdf to maven pom.xml
deletes unused files Folder and PhotoTag


can export one picture. its a start

it works, but is janky to use...

Added a folder selector to Export to PDF album.
For future work: Add a dialog with:
  * A list of images to export as PDF
  * Option to remove image from export list
  * Adda non static name to the album
  * etc...

Added a folder selector to Export to PDF album.
For future work: Add a dialog with:
  * A list of images to export as PDF
  * Option to remove image from export list
  * Adda non static name to the album
  * etc...

implements exclusions for Maven Shade plugin config


.gitignore .pdf files